requirement.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.0.8
django-cors-headers==3.11.0
djangorestframework==3.14.0
idna==3.4
pytz==2023.3
razorpay==1.3.0
requests==2.28.2
sqlparse==0.4.4
urllib3==1.26.15
